# perf: optimize `ScriptCoverage` normalization