… benchmark/fixtures/functions.json
… benchmark/fixtures/functions.ts
… test/__snapshots__/functions.json