- `vuejs/core` becomes 10s faster
- `vitejs/vite` becomes 2s faster (6s to 4s)
- Identical coverage on both projects before-after

```diff
-   ✓ normalized functions.json matches snapshot  38184ms
+   ✓ normalized functions.json matches snapshot 70ms
```

```diff
- Coverage remapping took 8.14s
+ Coverage remapping took 1.81s
```